test-embedding/test1/requirements.txt
python-dotenv
psycopg2-binary
google-generativeai
ollama
numpy
scikit-learn
pandas